+++ legit/scm.py
@@ -374,15 +374,12 @@ class SCMRepo(object):
 
         help = help.replace('Usage: legit', str('Usage: {0}'.format(crayons.black('legit', bold=True))))
 
-        help = help.replace('  branches', str(crayons.yellow('  branches', bold=True)))
-        help = help.replace('  install', str(crayons.magenta('  install', bold=True)))
-        help = help.replace('  uninstall', str(crayons.magenta('  uninstall', bold=True)))
-        help = help.replace('  publish', str(crayons.green('  publish', bold=True)))
-        help = help.replace('  unpublish', str(crayons.green('  unpublish', bold=True)))
-        help = help.replace('  settings', str(crayons.red('  settings', bold=True)))
         help = help.replace('  switch', str(crayons.green('  switch', bold=True)))
         help = help.replace('  sync', str(crayons.green('  sync', bold=True)))
+        help = help.replace('  publish', str(crayons.green('  publish', bold=True)))
+        help = help.replace('  unpublish', str(crayons.green('  unpublish', bold=True)))
         help = help.replace('  undo', str(crayons.green('  undo', bold=True)))
+        help = help.replace('  branches', str(crayons.yellow('  branches', bold=True)))
 
         additional_help = \
             """Usage Examples: